% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parmice.R
\name{parmice}
\alias{parmice}
\title{Parallel mice}
\usage{
parmice(..., paropts = list(.combine = ibind))
}
\arguments{
\item{...}{Arguments passed to \code{mice}}

\item{paropts}{A named list of arguments passed to \code{foreach}}
}
\value{
A \code{mids} object
}
\description{
Thin wrapper around mice to perform imputations in parallel
}
\examples{
library(mice)
library(doParallel)
library(parallel)
ncores <- 2
cl <- makeCluster(ncores)
clusterSetRNGStream(cl, 31313)
registerDoParallel(cl)
parmice(nhanes, m=5, paropts=list(".combine"=ibind))
}

