% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmice.R
\name{parmice}
\alias{parmice}
\title{Invoke mice function in parallel.}
\usage{
parmice(data, ncores = detectCores() - 1, m = 5,
  method = vector("character", length = ncol(data)), predictorMatrix = (1 -
  diag(1, ncol(data))), visitSequence = (1:ncol(data))[apply(is.na(data), 2,
  any)], form = vector("character", length = ncol(data)),
  post = vector("character", length = ncol(data)), defaultMethod = c("pmm",
  "logreg", "polyreg", "polr"), maxit = 5, diagnostics = TRUE,
  printFlag = TRUE, seed = NA, imputationMethod = NULL,
  defaultImputationMethod = NULL, data.init = NULL, ...)
}
\arguments{
\item{ncores}{Number of cores to be used.}
}
\value{
List of imputed datasets.
}
\description{
Invoke mice function in parallel.
}
\examples{
imp <-parmice(nhanes, ncores = 4)
plot(imp)
}
\seealso{
\code{\link[mice]{mice}}
}

